requirement.txt
openai
psycopg2-binary
python-dotenv
fastapi
jinja2
uvicorn
